generalized CommandManger to PubSubStack
